delete: 꿀팁 공유해요 불필요한 파일 삭제
