+++ app/src/main/java/com/umc/ttoklip/presentation/honeytip/adapter/HoneyTipAndQuestionVPA.kt
@@ -2,14 +2,14 @@ package com.umc.ttoklip.presentation.honeytip.adapter
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.umc.ttoklip.presentation.honeytip.HoneyTipFragment
+import com.umc.ttoklip.presentation.honeytip.InnerHoneyTipFragment
 import com.umc.ttoklip.presentation.honeytip.QuestionFragment
 
 class HoneyTipAndQuestionVPA(
     fragment: Fragment,
 ) : FragmentStateAdapter(fragment) {
     private var fragmentList = mutableListOf(
-        HoneyTipFragment(),
+        InnerHoneyTipFragment(),
         QuestionFragment()
     )
